refactor: Moving Liquidity component
