+++ apps/web/src/components/Merkl/MerklTag.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { Tag } from '@pancakeswap/uikit'
-import { useMerklInfo } from '../hooks/useMerkl'
+import { useMerklInfo } from '../../hooks/useMerkl'
 
 export function MerklTag({ poolAddress }: { poolAddress: string | null }) {
   const { t } = useTranslation()
