+++ apps/web/src/pages/liquidity/index.tsx
@@ -38,7 +38,7 @@ import { useMemo, useState } from 'react'
 import { V2PairCard } from 'views/AddLiquidityV3/components/V2PairCard'
 import { StablePairCard } from 'views/AddLiquidityV3/components/StablePairCard'
 import TransactionsModal from 'components/App/Transactions/TransactionsModal'
-import { LiquidityCardRow } from 'components/LiquidityCardRow'
+import { LiquidityCardRow } from 'views/AddLiquidity/components/LiquidityCardRow'
 import atomWithStorageWithErrorCatch from 'utils/atomWithStorageWithErrorCatch'
 import { useAtom } from 'jotai'
 import { V3SubgraphHealthIndicator } from 'components/SubgraphHealthIndicator'
