+++ apps/web/src/views/AddLiquidityV3/components/V2PairCard.tsx
@@ -3,7 +3,7 @@ import useTotalSupply from 'hooks/useTotalSupply'
 import { useTokensDeposited } from 'components/PositionCard'
 import { Tag } from '@pancakeswap/uikit'
 import { Pair } from '@pancakeswap/sdk'
-import { LiquidityCardRow } from 'components/LiquidityCardRow'
+import { LiquidityCardRow } from 'views/AddLiquidity/components/LiquidityCardRow'
 import { unwrappedToken } from 'utils/wrappedCurrency'
 import currencyId from 'utils/currencyId'
 import React from 'react'
